Refactor: Split triage and hubbub packages
